+++ src/main/java/com/carbonldp/config/ServicesConfig.java
@@ -11,10 +11,10 @@ import com.carbonldp.authentication.token.JWTAuthenticationService;
 import com.carbonldp.authentication.token.TokenService;
 import com.carbonldp.authorization.acl.ACLService;
 import com.carbonldp.authorization.acl.SesameACLService;
-import com.carbonldp.jobs.JobService;
-import com.carbonldp.jobs.SesameJobService;
-import com.carbonldp.jobs.SesameExecutionService;
 import com.carbonldp.jobs.ExecutionService;
+import com.carbonldp.jobs.JobService;
+import com.carbonldp.jobs.SesameExecutionService;
+import com.carbonldp.jobs.SesameJobService;
 import com.carbonldp.ldp.containers.ContainerService;
 import com.carbonldp.ldp.containers.SesameContainerService;
 import com.carbonldp.ldp.nonrdf.NonRDFSourceService;
@@ -95,8 +95,8 @@ public class ServicesConfig {
 	public JobService jobService() {return new SesameJobService(); }
 
 	@Bean
-	public ExecutionService executionService() {return new SesameExecutionService(); }
+	public BackupService backupService() {return new SesameBackupService(); }
 
 	@Bean
-	public BackupService backupService() {return new SesameBackupService(); }
+	ExecutionService executionService() {return new SesameExecutionService();}
 }