+++ src/main/java/com/carbonldp/jobs/JobService.java
@@ -14,4 +14,7 @@ public interface JobService {
 
 	@PreAuthorize( "hasPermission(#targetURI, 'READ')" )
 	public Job get( URI targetURI );
+
+	@PreAuthorize( "hasPermission(#jobURI, 'CREATE_CHILD')" )
+	public void createExecution( URI jobURI, Execution execution );
 }