+++ src/main/java/com/carbonldp/jobs/JobsExecutor.java
@@ -5,8 +5,10 @@ import com.carbonldp.apps.AppRepository;
 import com.carbonldp.exceptions.CarbonNoStackTraceRuntimeException;
 import com.carbonldp.exceptions.JobException;
 import com.carbonldp.models.Infraction;
+import com.carbonldp.apps.context.RunInPlatformContext;
+import com.carbonldp.authorization.Platform;
+import com.carbonldp.authorization.RunWith;
 import com.carbonldp.spring.TransactionWrapper;
-import org.openrdf.model.URI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,66 +20,58 @@ import java.util.List;
  * @author JorgeEspinosa
  * @since _version_
  */
+
 public class JobsExecutor {
-	protected final Logger LOG = LoggerFactory.getLogger( this.getClass() );
 	private final List<TypedJobExecutor> typedJobs;
-	private AppRepository appRepository;
 	private JobService jobService;
+	private ExecutionService executionService;
 	private TransactionWrapper transactionWrapper;
-	private ExecutionRepository executionRepository;
 
 	public JobsExecutor( List<TypedJobExecutor> typedJobs ) {
 		this.typedJobs = typedJobs;
 	}
 
 	@Async
-	public void execute( Execution execution ) {
-		LOG.debug( "Running execution " + Thread.currentThread().getName(), Thread.currentThread().getName() );
+	public void execute( App app, Execution execution ) {
+		transactionWrapper.runWithSystemPermissionsInPlatformContext( () -> manageExecution( app, execution ) );
+	}
 
-		executionRepository.changeExecutionStatus( execution.getURI(), ExecutionDescription.Status.RUNNING );
-		Job job = transactionWrapper.runWithSystemPermissionsInPlatformContext( () -> jobService.get( execution.getJobURI() ) );
+	private void manageExecution( App app, Execution execution ) {
+
+		executionService.changeExecutionStatus( execution.getURI(), ExecutionDescription.Status.RUNNING );
+		Job job = jobService.get( execution.getJobURI() );
 		JobDescription.Type type = JobFactory.getInstance().getJobType( job );
 		boolean hasErrors = false;
 
-		LOG.debug( "Running execution " + Thread.currentThread().getName() + " Job " + job.getSubject(), Thread.currentThread().getName() );
-
 		try {
-			transactionWrapper.runWithSystemPermissionsInPlatformContext( () -> getTypedRepository( type ).execute( job, execution ) );
+			getTypedRepository( type ).execute( job, execution );
 		} catch ( CarbonNoStackTraceRuntimeException e ) {
-			executionRepository.changeExecutionStatus( execution.getURI(), ExecutionDescription.Status.ERROR );
-			executionRepository.addErrorDescription( execution.getURI(), e.getMessage() );
+			executionService.changeExecutionStatus( execution.getURI(), ExecutionDescription.Status.ERROR );
+			executionService.addErrorDescription( execution.getURI(), e.getMessage() );
 			hasErrors = true;
 		} catch ( Exception e ) {
-			executionRepository.changeExecutionStatus( execution.getURI(), ExecutionDescription.Status.UNKNOWN );
+			executionService.changeExecutionStatus( execution.getURI(), ExecutionDescription.Status.UNKNOWN );
 			hasErrors = true;
 		}
-		if ( ! hasErrors ) executionRepository.changeExecutionStatus( execution.getURI(), ExecutionDescription.Status.FINISHED );
-		dequeueJobsExecutionQueue( job.getAppRelated() );
-
-		LOG.debug( "Ending execution " + Thread.currentThread().getName(), Thread.currentThread().getName() );
+		if ( ! hasErrors ) executionService.changeExecutionStatus( execution.getURI(), ExecutionDescription.Status.FINISHED );
+		executionService.dequeue( job.getURI( JobDescription.Property.EXECUTION_QUEUE_LOCATION.getURI() ) );
 	}
 
-	public TypedJobExecutor getTypedRepository( JobDescription.Type jobType ) {
+	private TypedJobExecutor getTypedRepository( JobDescription.Type jobType ) {
 		for ( TypedJobExecutor job : typedJobs ) {
 			if ( job.supports( jobType ) ) return job;
 		}
 		throw new IllegalArgumentException( "The jobType provided isn't supported" );
 	}
 
-	private void dequeueJobsExecutionQueue( URI appURI ) {
-		App app = appRepository.get( appURI );
-		appRepository.dequeueJobsExecutionQueue( app );
-	}
-
-	@Autowired
-	public void setAppRepository( AppRepository appRepository ) {this.appRepository = appRepository; }
-
 	@Autowired
 	public void setJobService( JobService jobService ) {this.jobService = jobService;}
 
 	@Autowired
-	public void setTransactionWrapper( TransactionWrapper transactionWrapper ) { this.transactionWrapper = transactionWrapper; }
+	public void setExecutionService( ExecutionService executionService ) { this.executionService = executionService; }
 
 	@Autowired
-	public void setExecutionRepository( ExecutionRepository executionRepository ) { this.executionRepository = executionRepository; }
+	public void setTransactionWrapper( TransactionWrapper transactionWrapper ) {
+		this.transactionWrapper = transactionWrapper;
+	}
 }