+++ src/main/java/com/carbonldp/jobs/web/JobsController.java
@@ -1,5 +1,6 @@
 package com.carbonldp.jobs.web;
 
+import com.carbonldp.Consts;
 import com.carbonldp.descriptions.APIPreferences;
 import com.carbonldp.rdf.RDFDocument;
 import com.carbonldp.web.AbstractController;
@@ -21,15 +22,26 @@ import javax.servlet.http.HttpServletResponse;
 @Controller
 @RequestMapping( value = "/platform/apps/*/jobs/" )
 public class JobsController extends AbstractController {
-	private JobPOSTHandler postRequestHandler;
+	private JobsPOSTHandler postRequestHandler;
 
 	@InteractionModel( value = {APIPreferences.InteractionModel.CONTAINER}, handlesDefault = true )
-	@RequestMapping( method = RequestMethod.POST )
+	@RequestMapping( method = RequestMethod.POST, consumes = {
+		Consts.RDFMediaTypes.TURTLE,
+		Consts.RDFMediaTypes.JSON_LD,
+		Consts.RDFMediaTypes.JSON_RDF,
+		Consts.RDFMediaTypes.XML_RDF,
+		Consts.RDFMediaTypes.TRIG,
+		Consts.RDFMediaTypes.N_TRIPLES,
+		Consts.RDFMediaTypes.N3,
+		Consts.RDFMediaTypes.TRIX,
+		Consts.RDFMediaTypes.BINARY,
+		Consts.RDFMediaTypes.N_QUADS
+	} )
 	public ResponseEntity<Object> createJob( @RequestBody RDFDocument requestDocument, HttpServletRequest request, HttpServletResponse response ) {
 		return postRequestHandler.handleRequest( requestDocument, request, response );
 	}
 
 	@Autowired
-	public void setPostRequestHandler( JobPOSTHandler postRequestHandler ) {this.postRequestHandler = postRequestHandler;}
+	public void setPostRequestHandler( JobsPOSTHandler postRequestHandler ) {this.postRequestHandler = postRequestHandler;}
 
 }