+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# 0.29.3 (2016/03/22)
+* Fixed validation when registering an Agent with an invalid document
+
+# 0.29.2 (2016/03/22)
+* Fixed SPARQL Query support for predefined documents (apps/, agents/, etc.)
+
+# 0.29.1 (2016/03/22)
+* Added support for the following Accept media types when executing SPARQL CONSTRUCT/DESCRIBE queries
+    - `application/trig`
+    - `application/n-triples`
+    - `text/n3`
+    - `application/trix`
+    - `application/x-binary-rdf`
+    - `application/n-quads`
+
+# 0.29.0 (2016/03/18)
+* Anonymous agents can now be added as a subject to ACEs
+
+# 0.28.5 (2016/03/18)
+* App cors filter won't trigger when asking an app document in the platform repository
+
 # 0.28.4 (2016/03/18)
 * Enable app role authentication in platform documents related to that app